Feat: Added Uniswap Swap & Liquidity Func Inside of LiquidityManager.sol Contract

+++ src/interfaces/IComet.sol
@@ -32,137 +32,64 @@ abstract contract CometMainInterface {
     error TransferOutFailed();
     error Unauthorized();
 
-    event Supply(address indexed from, address indexed dst, uint amount);
-    event Transfer(address indexed from, address indexed to, uint amount);
-    event Withdraw(address indexed src, address indexed to, uint amount);
+    event Supply(address indexed from, address indexed dst, uint256 amount);
+    event Transfer(address indexed from, address indexed to, uint256 amount);
+    event Withdraw(address indexed src, address indexed to, uint256 amount);
 
-    event SupplyCollateral(
-        address indexed from,
-        address indexed dst,
-        address indexed asset,
-        uint amount
-    );
-    event TransferCollateral(
-        address indexed from,
-        address indexed to,
-        address indexed asset,
-        uint amount
-    );
-    event WithdrawCollateral(
-        address indexed src,
-        address indexed to,
-        address indexed asset,
-        uint amount
-    );
+    event SupplyCollateral(address indexed from, address indexed dst, address indexed asset, uint256 amount);
+    event TransferCollateral(address indexed from, address indexed to, address indexed asset, uint256 amount);
+    event WithdrawCollateral(address indexed src, address indexed to, address indexed asset, uint256 amount);
 
     /// @notice Event emitted when a borrow position is absorbed by the protocol
-    event AbsorbDebt(
-        address indexed absorber,
-        address indexed borrower,
-        uint basePaidOut,
-        uint usdValue
-    );
+    event AbsorbDebt(address indexed absorber, address indexed borrower, uint256 basePaidOut, uint256 usdValue);
 
     /// @notice Event emitted when a user's collateral is absorbed by the protocol
     event AbsorbCollateral(
         address indexed absorber,
         address indexed borrower,
         address indexed asset,
-        uint collateralAbsorbed,
-        uint usdValue
+        uint256 collateralAbsorbed,
+        uint256 usdValue
     );
 
     /// @notice Event emitted when a collateral asset is purchased from the protocol
-    event BuyCollateral(
-        address indexed buyer,
-        address indexed asset,
-        uint baseAmount,
-        uint collateralAmount
-    );
+    event BuyCollateral(address indexed buyer, address indexed asset, uint256 baseAmount, uint256 collateralAmount);
 
     /// @notice Event emitted when an action is paused/unpaused
-    event PauseAction(
-        bool supplyPaused,
-        bool transferPaused,
-        bool withdrawPaused,
-        bool absorbPaused,
-        bool buyPaused
-    );
+    event PauseAction(bool supplyPaused, bool transferPaused, bool withdrawPaused, bool absorbPaused, bool buyPaused);
 
     /// @notice Event emitted when reserves are withdrawn by the governor
-    event WithdrawReserves(address indexed to, uint amount);
+    event WithdrawReserves(address indexed to, uint256 amount);
 
-    function supply(address asset, uint amount) external virtual;
+    function supply(address asset, uint256 amount) external virtual;
 
-    function supplyTo(address dst, address asset, uint amount) external virtual;
+    function supplyTo(address dst, address asset, uint256 amount) external virtual;
 
-    function supplyFrom(
-        address from,
-        address dst,
-        address asset,
-        uint amount
-    ) external virtual;
+    function supplyFrom(address from, address dst, address asset, uint256 amount) external virtual;
 
-    function transfer(address dst, uint amount) external virtual returns (bool);
+    function transfer(address dst, uint256 amount) external virtual returns (bool);
 
-    function transferFrom(
-        address src,
-        address dst,
-        uint amount
-    ) external virtual returns (bool);
+    function transferFrom(address src, address dst, uint256 amount) external virtual returns (bool);
 
-    function transferAsset(
-        address dst,
-        address asset,
-        uint amount
-    ) external virtual;
+    function transferAsset(address dst, address asset, uint256 amount) external virtual;
 
-    function transferAssetFrom(
-        address src,
-        address dst,
-        address asset,
-        uint amount
-    ) external virtual;
+    function transferAssetFrom(address src, address dst, address asset, uint256 amount) external virtual;
 
-    function withdraw(address asset, uint amount) external virtual;
+    function withdraw(address asset, uint256 amount) external virtual;
 
-    function withdrawTo(
-        address to,
-        address asset,
-        uint amount
-    ) external virtual;
+    function withdrawTo(address to, address asset, uint256 amount) external virtual;
 
-    function withdrawFrom(
-        address src,
-        address to,
-        address asset,
-        uint amount
-    ) external virtual;
+    function withdrawFrom(address src, address to, address asset, uint256 amount) external virtual;
 
-    function approveThis(
-        address manager,
-        address asset,
-        uint amount
-    ) external virtual;
+    function approveThis(address manager, address asset, uint256 amount) external virtual;
 
-    function withdrawReserves(address to, uint amount) external virtual;
+    function withdrawReserves(address to, uint256 amount) external virtual;
 
-    function absorb(
-        address absorber,
-        address[] calldata accounts
-    ) external virtual;
+    function absorb(address absorber, address[] calldata accounts) external virtual;
 
-    function buyCollateral(
-        address asset,
-        uint minAmount,
-        uint baseAmount,
-        address recipient
-    ) external virtual;
+    function buyCollateral(address asset, uint256 minAmount, uint256 baseAmount, address recipient) external virtual;
 
-    function quoteCollateral(
-        address asset,
-        uint baseAmount
-    ) public view virtual returns (uint);
+    function quoteCollateral(address asset, uint256 baseAmount) public view virtual returns (uint256);
 
     // function getAssetInfo(
     //     uint8 i
@@ -172,17 +99,13 @@ abstract contract CometMainInterface {
     //     address asset
     // ) public view virtual returns (AssetInfo memory);
 
-    function getCollateralReserves(
-        address asset
-    ) public view virtual returns (uint);
+    function getCollateralReserves(address asset) public view virtual returns (uint256);
 
-    function getReserves() public view virtual returns (int);
+    function getReserves() public view virtual returns (int256);
 
-    function getPrice(address priceFeed) public view virtual returns (uint);
+    function getPrice(address priceFeed) public view virtual returns (uint256);
 
-    function isBorrowCollateralized(
-        address account
-    ) public view virtual returns (bool);
+    function isBorrowCollateralized(address account) public view virtual returns (bool);
 
     function isLiquidatable(address account) public view virtual returns (bool);
 
@@ -192,17 +115,11 @@ abstract contract CometMainInterface {
 
     function balanceOf(address owner) public view virtual returns (uint256);
 
-    function borrowBalanceOf(
-        address account
-    ) public view virtual returns (uint256);
+    function borrowBalanceOf(address account) public view virtual returns (uint256);
 
-    function pause(
-        bool supplyPaused,
-        bool transferPaused,
-        bool withdrawPaused,
-        bool absorbPaused,
-        bool buyPaused
-    ) external virtual;
+    function pause(bool supplyPaused, bool transferPaused, bool withdrawPaused, bool absorbPaused, bool buyPaused)
+        external
+        virtual;
 
     function isSupplyPaused() public view virtual returns (bool);
 
@@ -216,15 +133,11 @@ abstract contract CometMainInterface {
 
     function accrueAccount(address account) external virtual;
 
-    function getSupplyRate(
-        uint utilization
-    ) public view virtual returns (uint64);
+    function getSupplyRate(uint256 utilization) public view virtual returns (uint64);
 
-    function getBorrowRate(
-        uint utilization
-    ) public view virtual returns (uint64);
+    function getBorrowRate(uint256 utilization) public view virtual returns (uint64);
 
-    function getUtilization() public view virtual returns (uint);
+    function getUtilization() public view virtual returns (uint256);
 
     function governor() external view virtual returns (address);
 
@@ -237,76 +150,52 @@ abstract contract CometMainInterface {
     function extensionDelegate() external view virtual returns (address);
 
     /// @dev uint64
-    function supplyKink() external view virtual returns (uint);
+    function supplyKink() external view virtual returns (uint256);
 
     /// @dev uint64
-    function supplyPerSecondInterestRateSlopeLow()
-        external
-        view
-        virtual
-        returns (uint);
+    function supplyPerSecondInterestRateSlopeLow() external view virtual returns (uint256);
 
     /// @dev uint64
-    function supplyPerSecondInterestRateSlopeHigh()
-        external
-        view
-        virtual
-        returns (uint);
+    function supplyPerSecondInterestRateSlopeHigh() external view virtual returns (uint256);
 
     /// @dev uint64
-    function supplyPerSecondInterestRateBase()
-        external
-        view
-        virtual
-        returns (uint);
+    function supplyPerSecondInterestRateBase() external view virtual returns (uint256);
 
     /// @dev uint64
-    function borrowKink() external view virtual returns (uint);
+    function borrowKink() external view virtual returns (uint256);
 
     /// @dev uint64
-    function borrowPerSecondInterestRateSlopeLow()
-        external
-        view
-        virtual
-        returns (uint);
+    function borrowPerSecondInterestRateSlopeLow() external view virtual returns (uint256);
 
     /// @dev uint64
-    function borrowPerSecondInterestRateSlopeHigh()
-        external
-        view
-        virtual
-        returns (uint);
+    function borrowPerSecondInterestRateSlopeHigh() external view virtual returns (uint256);
 
     /// @dev uint64
-    function borrowPerSecondInterestRateBase()
-        external
-        view
-        virtual
-        returns (uint);
+    function borrowPerSecondInterestRateBase() external view virtual returns (uint256);
 
     /// @dev uint64
-    function storeFrontPriceFactor() external view virtual returns (uint);
+    function storeFrontPriceFactor() external view virtual returns (uint256);
 
     /// @dev uint64
-    function baseScale() external view virtual returns (uint);
+    function baseScale() external view virtual returns (uint256);
 
     /// @dev uint64
-    function trackingIndexScale() external view virtual returns (uint);
+    function trackingIndexScale() external view virtual returns (uint256);
 
     /// @dev uint64
-    function baseTrackingSupplySpeed() external view virtual returns (uint);
+    function baseTrackingSupplySpeed() external view virtual returns (uint256);
 
     /// @dev uint64
-    function baseTrackingBorrowSpeed() external view virtual returns (uint);
+    function baseTrackingBorrowSpeed() external view virtual returns (uint256);
 
     /// @dev uint104
-    function baseMinForRewards() external view virtual returns (uint);
+    function baseMinForRewards() external view virtual returns (uint256);
 
     /// @dev uint104
-    function baseBorrowMin() external view virtual returns (uint);
+    function baseBorrowMin() external view virtual returns (uint256);
 
     /// @dev uint104
-    function targetReserves() external view virtual returns (uint);
+    function targetReserves() external view virtual returns (uint256);
 
     function numAssets() external view virtual returns (uint8);
